+++ sdk/cosmos/azure-cosmos/samples/cosmos_opentelemetry_example.py
@@ -28,8 +28,8 @@ trace.set_tracer_provider(provider)
 settings.tracing_implementation = OpenTelemetrySpan
 
 # --- Cosmos DB setup ---
-COSMOS_ENDPOINT = "https://localhost:8081"
-COSMOS_KEY = "C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw=="
+COSMOS_ENDPOINT = "host"
+COSMOS_KEY = "key"
 
 # --- Default behavior example ---
 def run_default_behavior():